LayoutTests/imported/w3c/web-platform-tests/css/css-flexbox/align-items-baseline-column-horz-expected.txt

line1
line2
line1
line2
line1
line2
line1
line2

PASS #target > div 1
PASS #target > div 2
FAIL #target > div 3 assert_equals:
<div data-offset-x="37"></div>
offsetLeft expected 37 but got 67
PASS #target > div 4
PASS #target > div 5
